fix tests. work around circular dependencies in sync rules package
