+++ packages/sync-rules/src/sql_support.ts
@@ -1,3 +1,8 @@
+import { SelectFromStatement } from 'pgsql-ast-parser';
+import { SqlRuleError } from './errors.js';
+import { ExpressionType } from './ExpressionType.js';
+import { MATCH_CONST_FALSE, MATCH_CONST_TRUE } from './sql_filters.js';
+import { evaluateOperator, getOperatorReturnType } from './sql_functions.js';
 import {
   ClauseError,
   CompiledClause,
@@ -6,16 +11,11 @@ import {
   ParameterMatchClause,
   ParameterValueClause,
   QueryParameters,
-  SqliteValue,
   RowValueClause,
+  SqliteValue,
   StaticValueClause,
   TrueIfParametersMatch
 } from './types.js';
-import { MATCH_CONST_FALSE, MATCH_CONST_TRUE } from './sql_filters.js';
-import { SqlFunction, evaluateOperator, getOperatorReturnType } from './sql_functions.js';
-import { SelectFromStatement } from 'pgsql-ast-parser';
-import { SqlRuleError } from './errors.js';
-import { ExpressionType } from './ExpressionType.js';
 
 export function isParameterMatchClause(clause: CompiledClause): clause is ParameterMatchClause {
   return Array.isArray((clause as ParameterMatchClause).inputParameters);
@@ -74,18 +74,6 @@ export function compileStaticOperator(op: string, left: RowValueClause, right: R
         name: '?',
         type
       };
-    }
-  };
-}
-
-export function getOperatorFunction(op: string): SqlFunction {
-  return {
-    debugName: `operator${op}`,
-    call(...args: SqliteValue[]) {
-      return evaluateOperator(op, args[0], args[1]);
-    },
-    getReturnType(args) {
-      return getOperatorReturnType(op, args[0], args[1]);
     }
   };
 }
